ZavrsniAplikacija

+++ Csharp/Kuca/KucaVjezbanje/ZavrsniAplikacija/Model/Ture.cs
@@ -9,8 +9,8 @@ namespace KucaVjezbanje.ZavrsniAplikacija.Model
     internal class Ture
     {
         public int? Tura_ID { get; set; }
-        public int? Prijedeni_Km { get; set; }
-        public int? Udaljenost { get; set; }
+        public double? Prijedeni_Km { get; set; }
+        public double? Udaljenost { get; set; }
         public DateTime? Datum { get; set; }
         public double? Potrosnja_Goriva { get; set; }
         public int? Kamion_ID { get; set; }
